Converted Enma to a static library & various mantainance

+++ Enma/src/Platform/Windows/WindowsWindow.cpp
@@ -136,7 +136,7 @@ namespace Enma
 		{
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
 
-			MouseScrolledEvent event(xOffset, yOffset);
+			MouseScrolledEvent event((float)xOffset, (float)yOffset);
 			data.EventCallback(event);
 		});
